Better report docker command errors.
This ensures that if a docker command (either docker or docker-compose) fails, the contents of stderr is reported as a VSTS error so that it shows up in the build/release summary.

Also provide better error messages when no Dockerfile or docker-compose.yml files are found with the provided glob patterns.

Finally, fix an issue with docker-compose run action where it doesn't do docker-compose down if the run failed.